load gpt2 model as type dtype; TODO: fix nan loss

+++ gpt_2_simple/src/sample.py
@@ -68,7 +68,7 @@ def sample_sequence(*, hparams, length, start_token=None,
 
         def body(past, prev, output):
             next_outputs = step(hparams, prev[:, tf.newaxis], past=past)
-            logits = next_outputs['logits'][:, -1, :] / tf.cast(temperature, tf.float32)
+            logits = next_outputs['logits'][:, -1, :] / tf.cast(temperature, hparams.dtype)
             if top_p > 0.0:
                 logits = top_p_logits(logits, p=top_p)
             else:
